Setting up for zql parsing
