+++ src/main/java/edu/uiuc/zenvisage/data/remotedb/VisualComponentList.java
@@ -7,9 +7,6 @@ import java.util.List;
 
 public class VisualComponentList {
 
-	private String ztype;
-	private String xType;
-    private String ytype;
 	public ArrayList<VisualComponent> visualComponentList;
     public HashMap<String,VisualComponent> ZToVisualComponents = new HashMap<>();
 	
@@ -60,46 +57,5 @@ public class VisualComponentList {
 		return ret.toString();
 	}
 	
-	/**
-	 * @return the ztype
-	 */
-	public String getZtype() {
-		return ztype;
-	}
-
-	/**
-	 * @param ztype the ztype to set
-	 */
-	public void setZtype(String ztype) {
-		this.ztype = ztype;
-	}
-
-	/**
-	 * @return the xType
-	 */
-	public String getxType() {
-		return xType;
-	}
-
-	/**
-	 * @param xType the xType to set
-	 */
-	public void setxType(String xType) {
-		this.xType = xType;
-	}
-
-	/**
-	 * @return the ytype
-	 */
-	public String getYtype() {
-		return ytype;
-	}
-
-	/**
-	 * @param ytype the ytype to set
-	 */
-	public void setYtype(String ytype) {
-		this.ytype = ytype;
-	}
 
 }
